Make NetworkNodeExample Complete
The ipv4 layer must be named `ipv4` to ensure addresses are assigned to interfaces. It should also be an `Ipv4NetworkLayer` type so as to include routing tables and configurators.

Additionally, the radioIn must be specified as `@directIn` for connection to the radioMedium in a network.

Finally, the parameters must be configured correctly. For example it must be of type NetworkNode and the mobility module must be defined to the wireless interfaces